moved logic to Zendesk class
